+++ maven/plugin/xdb-maven-plugin/src/test/java/org/safris/maven/plugin/xdb/xde/SetFunctionTest.java
@@ -52,6 +52,6 @@ public class SetFunctionTest extends LoggableTest {
     Assert.assertEquals(BigInteger.valueOf(24367857008l), rows.nextEntity().get());
     Assert.assertEquals("White Plains", rows.nextEntity().get());
     Assert.assertEquals(classicmodels.Address.Country.AU, rows.nextEntity().get());
-    Assert.assertEquals(BigInteger.valueOf(21003), rows.nextEntity().get());
+    Assert.assertEquals(Long.valueOf(21003), rows.nextEntity().get());
   }
 }
